Build script added. JavaScript source code moved to srcjs folder.
The build script only minifies the source code of the template but
does not compile the Joshfire framework. The resulting code is
stored in app/js.

The app/srcjs folder has been added to the list of folders to ignore
when the template is imported.

For debug/development purpose, index files need to be adjusted to
link to "srcjs" instead of "js" in <script data-main="js"> tag.